pyproject.toml
[project]
name = "mp_back_common"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]

readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "aioinject (>=0.38.1,<0.39.0)",
    "elasticsearch (>=9.0.1,<10.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.group.dev.dependencies]
ruff = "^0.11.11"


[tool.poetry.group.isort.dependencies]
ruff = "^0.11.11"
pyclean = "^3.1.0"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.9
target-version = "py312"

[lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"